Avoid cyclic module dependencies
